handling client entering and leaving store

+++ projservice/storegoservice/src/main/java/com/storego/storegoservice/model/Cart.java
@@ -7,17 +7,17 @@ import java.util.Set;
 @Table(name = "cart")
 public class Cart {
 
-    private int nif;
+    private long nif;
     private Person client;
     private Set<CartProduct> cartproducts;
 
     public Cart() {}
 
     @Id
-    public int getNif() {
+    public long getNif() {
         return nif;
     }
-    public void setNif(int nif) {
+    public void setNif(long nif) {
         this.nif = nif;
     }
